+++ CryptoData/app/src/main/java/dev/shorthouse/cryptodata/data/source/remote/model/CoinDetailApiModel.kt
@@ -1,6 +1,7 @@
 package dev.shorthouse.cryptodata.data.source.remote.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class CoinDetailApiModel(
     val id: String,
@@ -8,19 +9,19 @@ data class CoinDetailApiModel(
     val symbol: String?,
     val image: String?,
     @SerializedName("current_price")
-    val currentPrice: Double?,
+    val currentPrice: BigDecimal?,
     @SerializedName("market_cap")
-    val marketCap: Double?,
+    val marketCap: BigDecimal?,
     @SerializedName("market_cap_rank")
     val marketCapRank: Int?,
     @SerializedName("circulating_supply")
-    val circulatingSupply: Double?,
+    val circulatingSupply: BigDecimal?,
     @SerializedName("atl")
-    val allTimeLow: Double?,
+    val allTimeLow: BigDecimal?,
     @SerializedName("atl_date")
     val allTimeLowDate: String?,
     @SerializedName("ath")
-    val allTimeHigh: Double?,
+    val allTimeHigh: BigDecimal?,
     @SerializedName("ath_date")
     val allTimeHighDate: String?
 )
