+++ CryptoData/app/src/main/java/dev/shorthouse/cryptodata/model/CoinChart.kt
@@ -1,7 +1,9 @@
 package dev.shorthouse.cryptodata.model
 
+import java.math.BigDecimal
+
 data class CoinChart(
-    val prices: List<Double>,
+    val prices: List<BigDecimal>,
     val minPrice: Price,
     val minPriceChangePercentage: Percentage,
     val maxPrice: Price,
